refactored chainio/clients module
